Create Randomswitching Array (USES EXTERNAL FILES)
Quite a simple array with randomised (but controlled) length and playback speed on a few sound files, which can be switched out for anything else. Use this one in live sets with calls to different harsh noise loops. This array setup can easily be tweaked for use in a lot of aleatoric music.